Powerups fixes and additions plus crushes
SPRITES, SKINS, DECORATE, DEFS, LANGUAGE & CREDITS
:new: Nitpicked some lumps
:new: Added new behavior for Berserk_ : now it works like the common
berserk even for Caleb, but it switches to Pitchfork. This poweup must
be set only for Coop/DM on maps
:new: Added 2 new powerups with Doom Beta sprites: the BeastForm for
Caleb and the UnholyBible_ for Doomguy. Set these poerups only for
SinglePlayer on maps and use them in conjunction with the Berserk_
pickup at same position
:new: Added LANGUAGE entries for bothe new powerups and updated credits
:new: Crushed recently added sprites again (saved a couple of kbs XD )
and moved Ice decals inside Models folder
:new: TESTMAP - Added both new powerups at the start, plus berserk
behind us
:new: BLM01 - Added both new powerups at the crypt on the courtyard with
proper scripts and flags as mentioned on FB chat